fix(@warp-ds/toast): remove window reference from toast api bundle, fixing server-side rendering
The toast API relies on a utility function to prevent browser APIs from being called in a
non-browser environment (eg. Node). The utility function identifies non-browser environments by
checking  if the window object is defined. The function was imported from a package that also
imported the classmap function from lit-html, which in itself contains a reference to the window
object on top-level of the module. This causes server-side rendering to fail as soon as the imported
modules are parsed and evaluated. To fix this issue, the utility function in question is moved to a
separate module that does not contain transitive references to the window object, and it can now
safely be used for server-side rendering.