refactor: generate modal style using unocss instead of relying on components.css
If Modal will be reused across frameworks in Warp, there shouldn't be a need to keep its classes in
a "shared" css repo. This a test to see how we could use UnoCSS to generate styles for each web
component the way it's designed. Note: we might need to consider importing resets using Eik here.