chore: accessibility refactor for screen readers
The calendar is rewritten as a focus-trapped modal to have it announced
(in my opinion) in a better way for screen readers. This is in line with
the example from the ARIA Author Practices Guide

https://www.w3.org/WAI/ARIA/apg/patterns/dialog-modal/examples/datepicker-dialog/

For the announcement to work correctly we can't conditionally render the
calendar, it must be present in the document. We toggle its presence
both visual and in the accessibility tree with display: none;