fix: fix interaction between steps and allowing values outside range
add custom step handling to fix the interaction between steps and
allowing values outside range. when `allowValuesOutsideRange` is enabled
the range slider goes from min-1 to max+1, making the step be off by
one. adding a custom stepper allows us to handle this case, and allows
the user to get close to the end, and then have the extra step to enable
"before" or "after" search